refactor: split large responsibility modules
