+++ src/contexts/evaluation/infrastructure/gh/fetch.rs
@@ -1,0 +1,31 @@
+use std::path::Path;
+
+use super::command::run_gh;
+use super::schema::{GhCompare, GhRepoView};
+
+/// GitHub Compare API でベースブランチとの差分コミット数を取得する。
+///
+/// `base_ref` / `head_ref` が空の場合は `Some(0)` を返す（追跡不要）。
+/// 失敗した場合は `None` を返す（呼び出し元が `SyncUnknown` として扱う）。
+pub(super) fn fetch_behind_by(base_ref: &str, head_ref: &str, cwd: Option<&Path>) -> Option<u64> {
+    if base_ref.is_empty() || head_ref.is_empty() {
+        return Some(0);
+    }
+
+    let name_with_owner = match run_gh(&["repo", "view", "--json", "nameWithOwner"], cwd) {
+        Ok(bytes) => match serde_json::from_slice::<GhRepoView>(&bytes) {
+            Ok(r) => r.name_with_owner,
+            Err(_) => return None,
+        },
+        Err(_) => return None,
+    };
+
+    let path = format!("repos/{name_with_owner}/compare/{base_ref}...{head_ref}");
+
+    match run_gh(&["api", &path], cwd) {
+        Ok(bytes) => serde_json::from_slice::<GhCompare>(&bytes)
+            .map(|c| c.behind_by)
+            .ok(),
+        Err(_) => None,
+    }
+}
