+++ tests/e2e/config/mod.rs
@@ -1,1 +1,3 @@
-mod config;
+mod edit;
+mod prompt_format;
+mod xdg;
