+++ tests/e2e/daemon/mod.rs
@@ -1,2 +1,4 @@
+mod concurrency;
 mod lifecycle;
+mod restart;
 mod watch;
